+++ server/api/index.js
@@ -2,7 +2,7 @@ const router = require('express').Router()
 
 router.use('/users', require('./users'))
 router.use('/products', require('./products'))
-router.use('/orders', require('./order'))
+router.use('/orders', require('./orders'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
