+++ server/db/models/product.spec.js
@@ -2,7 +2,7 @@ const {expect} = require('chai')
 const db = require('../index')
 const Product = db.model('product')
 
-describe('Product Model', () => {
+describe.only('Product Model', () => {
   let product
   before(() => db.sync({force: true}))
   beforeEach(() => {
